A lot of basic infrastructure, the start of physical memory access.
The idea with physical memory access is that we always ensure that we
can map and remap the (1GiB) page at address 0xffffffff80000000. When we
need to access physical memory, we remap that virtual address to the
right place, then do the access.

This is not efficient, and needs locks around it that currently are not
present. However, it's easy to understand, and efficient enough for
early boot.