Adds symbolicated backtraces.
This is currently complicated by compiler optimizations -- code paths
leading to a panic commonly get moved out-of-line, and the link address
then often points to the first byte of the next function.

This may just be an artifact of the small functions that are currently
in the kernel. Either way, they're _much_ better than nothing.