bootstub: Pass the range the kernel is in.
This will let us avoid treating those pages as allocatable memory, and
will let us ensure we _can_ treat the ones we want to as allocatable
later.

In particular, the initial page tables and initial stack should both be
reclaimable -- we make sure they're covered by the buddy allocator we
set up, so whenever they get freed, they can be added back to the buddy
system. (We'll do the same for the DeviceTree as well.)