Adds a random number generator API.
In the short term, this will let us randomly shuffle allocator free
lists and generate cookies to XOR-encode them, improving security and
increasing the odds heap overflow bugs are found.

In the slightly longer term, once we have timers and interrupts, we can
start putting real entropy into the pool, and this can be the RNG we
expose to userspace and use anywhere else in the kernel that needs
random number generation.