Aborted attempt to add USB Mass Storage mode to U-Boot.
This required adjusting the dwc2-otg driver in U-Boot; I guess the
reason why upstream hadn't done so is because it doesn't work.

I'm committing this as a checkpoint, before tearing out this attempt; if
someone else wants to pick this up later, this might make a good
starting point.

Signed-off-by: Nathan Ringo <me@remexre.com>