Adds preliminary support for the Milk-V Duo S. (#15)
* Adds Nix package definition for U-Boot for the Duo S.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Adds Nix package definition for OpenSBI for the Duo S.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Adds Nix package definition for FSBL for the Duo S.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Use unwrapped compilers for FSBL.

This gets loaded directly, without applying relocations or anything.
Normally this would be fine, but the hardening flags in the wrapped
compiler make the program use the GOT very early in boot (in bss_clear
in lib/cpu/riscv/bl2_entrypoint.S), which then crashes.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Aborted attempt to add USB Mass Storage mode to U-Boot.

This required adjusting the dwc2-otg driver in U-Boot; I guess the
reason why upstream hadn't done so is because it doesn't work.

I'm committing this as a checkpoint, before tearing out this attempt; if
someone else wants to pick this up later, this might make a good
starting point.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Adds a boot menu and tftpd boot.

This allows developers to upload a kernel to the board over TFTP, rather
than the other way around. This way, it does not require binding to a
privileged port on the developer machine.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Switch to use the (old, deprecated) sbi_console_putchar.

The Sophgo fork of OpenSBI only supports this method, sadly...
Long-term, we should tear this out and probe for DBCN anyway.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Patches U-Boot to load our kernel with the same ABI as QEMU.

In theory, we could package our kernel like how the Linux kernel gets
packaged, and use the bootm command instead. However, this would stop us
from using ELF tooling, so I'm putting it off for now.

Signed-off-by: Nathan Ringo <me@remexre.com>

* Makes GitHub actions build the dev image.

Signed-off-by: Nathan Ringo <me@remexre.com>

---------

Signed-off-by: Nathan Ringo <me@remexre.com>