Apply SPDX markings to more files.
This removes a few cases that I special-cased in the REUSE.toml because
I was too lazy to put the actual headers in. This isn't ideal, because
then people who copy those files are wondering how they pass the CI job
if they're not familiar with REUSE.toml.

Also fixes the README claiming that CI configs are CC0, when they're
actually GPL3.

Signed-off-by: Nathan Ringo <me@remexre.com>